extract Base* encoders and ScorexEncoding trait to scorex-util;
