+++ build.sbt
@@ -4,7 +4,7 @@ name := "scrypto"
 
 lazy val commonSettings = Seq(
   organization := "org.scorexfoundation",
-  version := "2.1.2",
+  version := "2.1.3-SNAPSHOT",
   scalaVersion := "2.12.5",
   licenses := Seq("CC0" -> url("https://creativecommons.org/publicdomain/zero/1.0/legalcode")),
   homepage := Some(url("https://github.com/input-output-hk/scrypto")),
@@ -28,7 +28,8 @@ libraryDependencies ++= Seq(
   "com.typesafe.scala-logging" %% "scala-logging" % "3.+",
   "ch.qos.logback" % "logback-classic" % "1.+",
   "org.whispersystems" % "curve25519-java" % "+",
-  "org.bouncycastle" % "bcprov-jdk15on" % "1.+"
+  "org.bouncycastle" % "bcprov-jdk15on" % "1.+",
+  "org.scorexfoundation" %% "scorex-util" % "0.1.1-SNAPSHOT"
 )
 
 libraryDependencies ++= Seq(
