+++ src/main/scala/scorex/crypto/authds/avltree/batch/ToStringHelper.scala
@@ -1,6 +1,7 @@
 package scorex.crypto.authds.avltree.batch
 
-import scorex.utils.{ScorexEncoding, ScryptoLogging}
+import scorex.utils.ScryptoLogging
+import scorex.util.ScorexEncoding
 
 trait ToStringHelper extends ScryptoLogging with ScorexEncoding {
   //Needed for debug (toString) only
